# docs/openapi.yaml
openapi: 3.0.3
info:
  title: iDME 工业 MiniApp API
  description: |
    华为 iDME 数据建模引擎 - 工业 MiniApp 接口文档。

    ## 快速开始

    ### 推荐创建顺序（依赖链）

    **Level 0（无依赖）**：Unit（计量单位）, BusinessPartner（往来单位）, EquipmentClassfication（设备分类）, PartClassfication（物料分类）

    **Level 1**：Department（部门）, Employee（员工）

    **Level 2**：EquipmentModel（设备型号）, Location（位置）

    **Level 3**：Equipment（设备）, Part（物料）

    **Level 4**：WorkingProcedure（工序）, WorkingPlan（工艺路线）

    **Level 5**：BOMItem（BOM项）, PlanProcedureLink（计划-工序关联）, ProcedureEquipmentLink（工序-设备关联）, ProcedurePartLink（工序-物料关联）, EquipmentSparePartLink（设备-备件关联）

    ### 注意事项
    - **引用字段**（如 `manufacturerName`, `category`）需要传入对应模块的 **ID**，不是文本名称
    - **Part** 和 **WorkingPlan** 是版本对象，`WorkingPlan.productPart` 引用 Part 时使用 `master.id` 而非实例 `id`
    - **枚举值**使用英文（如 `Standby`, `Active`, `Machinery`）
    - 所有接口的 example 字段已填充真实可用的示例ID，可直接复制使用

    ### 示例数据
    - 固定ID文档: `/docs/openapi-examples-ids.md`
    - 测试文件: `/src/test/java/com/tsukilc/idme/integration/OpenApiExamplesDataTest.java`

    ---

    - 管理界面: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/index.html
    - 健康检查: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services/v1/health

    认证: Basic Auth 用户名 sysadmin / 密码见项目说明。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://99.suyiiyii.top:8003/rdm_a1b52ff379ee46ed8928d7f2ceb908f6_app/services
    description: iDME SDK 服务
security:
  - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 用户名 sysadmin，密码见项目文档
  parameters:
    pageNum:
      name: pageNum
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: 每页大小
      schema:
        type: integer
        default: 20
    idPath:
      name: id
      in: path
      required: true
      description: 主键 ID
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        result:
          type: string
          example: SUCCESS
        data: {}
        errors:
          type: array
          items:
            type: string
      description: 统一响应包装
...
    PartnerRef:
      type: object
      description: 往来单位引用
      properties:
        id: { type: string }
        partnerName: { type: string }
    PartRef:
      type: object
      description: 物料引用（含名称）
      properties:
        id: { type: string }
        partName: { type: string }
    PartRefWithNumber:
      type: object
      description: 物料引用（含编号、名称）
      properties:
        id: { type: string }
        partNumber: { type: string }
        partName: { type: string }
...
    BOMTreeNode:
      type: object
      description: BOM 树节点
      properties:
        part: { $ref: '#/components/schemas/PartRefWithNumber' }
        findNumber: { type: integer }
        quantity: { type: number }
        uom: { type: string }
        children:
          type: array
          items: { $ref: '#/components/schemas/BOMTreeNode' }
...
    EquipmentDetail:
      type: object
      description: 设备详情（含关联对象）
      properties:
        id: { type: string }
        equipmentCode: { type: string }
        equipmentName: { type: string }
        manufacturerName:
          oneOf:
            - { type: string }
            - { $ref: '#/components/schemas/PartnerRef' }
        brand: { type: string }
        modelSpec: { type: string }
        status: { type: string, enum: [运行, 待机, 维修, 停机, 报废] }
        techParams: { type: object, additionalProperties: true }
        createTime: { type: string, format: date-time }
        lastModifiedTime: { type: string, format: date-time }
    EquipmentListItem:
      type: object
      description: 设备列表项
      properties:
        id: { type: string }
        equipmentCode: { type: string }
        equipmentName: { type: string }
        status: { type: string }
        locationText: { type: string }
...
    BOMItemDetail:
      type: object
      description: BOM 项详情（含父件、子件引用）
      properties:
        id: { type: string }
        parentPart: { $ref: '#/components/schemas/PartRef' }
        childPart: { $ref: '#/components/schemas/PartRef' }
        quantity: { type: number }
        uom: { type: string }
        findNumber: { type: integer }
    BOMTreeRoot:
      type: object
      description: BOM 树根节点
      properties:
        part: { $ref: '#/components/schemas/PartRefWithNumber' }
        children:
          type: array
          items: { $ref: '#/components/schemas/BOMTreeNode' }
...
    EquipmentModel:
      type: object
      properties:
        id: { type: string }
        modelCode: { type: string }
        modelName: { type: string }
        manufacturer: { oneOf: [{ type: string }, { $ref: '#/components/schemas/PartnerRef' }] }
        brand: { type: string }
        modelSpec: { type: string }
        category: { type: string }
        defaultTechParams: { type: object }
        remarks: { type: string }
...
    EquipmentCreate:
      type: object
      required: [equipmentCode, equipmentName, manufacturerName]
      properties:
        equipmentCode:
          type: string
          description: 设备编码（唯一）
          example: "EQ-PROD-DEMO"
        equipmentName:
          type: string
          description: 设备名称
          example: "CNC加工中心A"
        manufacturerName:
          type: string
          description: 生产厂家ID（-> BusinessPartner模块，partnerType=Machinery）
          example: "2000000001"
        brand:
          type: string
          description: 品牌
          example: "华为"
        modelSpec:
          type: string
          description: 型号规格
        equipmentModelRef:
          type: string
          description: 设备型号ID（-> EquipmentModel模块，SDK必填）
          example: "5000000001"
        supplierName:
          type: string
          description: 供应商ID（-> BusinessPartner模块，SDK必填）
          example: "2000000002"
        productionDate:
          type: string
          format: date
          description: 生产日期
        serviceLifeYears:
          type: integer
          description: 使用年限
        depreciationMethod:
          type: string
          enum: [NoDepreciation, StraightLine, Double, SumOfYear]
          description: "折旧方式：不折旧/直线法/双倍余额递减法/年数总和法"
          example: "StraightLine"
        locationRef:
          type: string
          description: 位置ID（-> Location模块，SDK必填）
          example: "5000000002"
        locationText:
          type: string
          description: 位置文本描述
        status:
          type: string
          enum: [Standby, InOperation, UnderMaintenance, Shutdown, Scrap]
          description: "设备状态：待机/运行中/维修中/停机/报废"
          example: "Standby"
        serialNumber:
          type: string
          description: 序列号
          example: "SN-PROD-001"
        category:
          type: string
          description: 设备分类ID（-> EquipmentClassfication模块）
          example: "3000000001"
        techParams:
          type: object
          description: 技术参数（JSON扩展）
          example: {"精度": "0.001mm", "功率": "7.5kW"}
        remarks:
          type: string
          description: 备注
          example: "示例设备"
...
    BOMItem:
      type: object
      properties:
        id: { type: string }
        parentPart: { type: object }
        childPart: { type: object }
        quantity: { type: number }
        uom: { type: string }
        findNumber: { type: integer }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: string, format: date }
        remarks: { type: string }
    BOMItemCreate:
      type: object
      required: [parentPart, childPart, quantity]
      properties:
        parentPart:
          type: string
          description: 父件ID（-> Part模块）
          example: "7000000001"
        childPart:
          type: string
          description: 子件ID（-> Part模块，⚠️不能与parentPart相同，否则触发循环引用检测）
          example: "7000000002"
        quantity:
          type: number
          description: 用量
          example: 2.0
        uom:
          type: string
          description: 计量单位ID（-> Unit模块）
          example: "1000000001"
        findNumber:
          type: integer
          description: 查找号
          example: 1
        effectiveFrom:
          type: string
          format: date
          description: 生效开始日期
          example: "2024-01-01"
        effectiveTo:
          type: string
          format: date
          description: 生效结束日期
        remarks:
          type: string
          description: 备注
...
    EquipmentModelCreate:
      type: object
      required: [modelCode, modelName]
      properties:
        modelCode:
          type: string
          description: 型号编码
          example: "MODEL-DEMO"
        modelName:
          type: string
          description: 型号名称
          example: "CNC-5000"
        manufacturer:
          type: string
          description: 生产厂家ID（-> BusinessPartner模块，partnerType=Machinery）
          example: "2000000001"
        brand:
          type: string
          description: 品牌
          example: "华为"
        modelSpec:
          type: string
          description: 型号规格
        category:
          type: string
          description: 设备分类ID（-> EquipmentClassfication模块）
          example: "3000000001"
        defaultTechParams:
          type: object
          description: 默认技术参数（JSON扩展）
        remarks:
          type: string
          description: 备注
...
paths:
...
  /api/equipment:
    get:
      tags: [设备管理]
      summary: 分页查询设备列表
      operationId: listEquipment
      parameters:
        - $ref: '#/components/parameters/pageNum'
        - $ref: '#/components/parameters/pageSize'
        - name: equipmentName
          in: query
          schema: { type: string }
          description: 设备名称（模糊）
        - name: equipmentCode
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          pageNum: { type: integer }
                          pageSize: { type: integer }
                          records:
                            type: array
                            items: { $ref: '#/components/schemas/EquipmentListItem' }
    post:
      tags: [设备管理]
      summary: 创建设备
      operationId: createEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EquipmentDetail'
  /api/equipment/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [设备管理]
      summary: 查询设备详情
      operationId: getEquipment
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EquipmentDetail'
    put:
      tags: [设备管理]
      summary: 更新设备
      operationId: updateEquipment
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EquipmentCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [设备管理]
      summary: 删除设备
      operationId: deleteEquipment
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/equipment/{id}/tech-params:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [设备管理]
      summary: 查询设备技术参数
      operationId: getEquipmentTechParams
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: true
                        description: 技术参数 JSON 对象，如 maxSpeed、power、precision 等
    put:
      tags: [设备管理]
      summary: 更新设备技术参数
      operationId: updateEquipmentTechParams
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/bom-item:
    post:
      tags: [BOM]
      summary: 创建 BOM 项
      operationId: createBOMItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BOMItemCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMItemDetail'
  /api/bom-item/batch:
    post:
      tags: [BOM]
      summary: 批量创建 BOM 项
      operationId: batchCreateBOMItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/BOMItemCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/bom-item/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [BOM]
      summary: 查询 BOM 项详情
      operationId: getBOMItem
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMItemDetail'
    put:
      tags: [BOM]
      summary: 更新 BOM 项
      operationId: updateBOMItem
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BOMItemCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [BOM]
      summary: 删除 BOM 项
      operationId: deleteBOMItem
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/bom-item/tree/{partId}:
    get:
      tags: [BOM]
      summary: 查询物料 BOM 结构（树形）
      operationId: getBOMTree
      parameters:
        - name: partId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMTreeRoot'
  /api/bom-item/where-used/{partId}:
    get:
      tags: [BOM]
      summary: 查询物料被哪些父项使用
      operationId: getBOMWhereUsed
      parameters:
        - name: partId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/BOMItem' }
...
  /api/equipment-model:
    get:
      tags: [设备机型]
      summary: 分页查询设备机型列表
      operationId: listEquipmentModel
      parameters:
        - $ref: '#/components/parameters/pageNum'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          pageNum: { type: integer }
                          pageSize: { type: integer }
                          records:
                            type: array
                            items: { $ref: '#/components/schemas/EquipmentModel' }
    post:
      tags: [设备机型]
      summary: 创建设备机型
      operationId: createEquipmentModel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EquipmentModelCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EquipmentModel'
  /api/equipment-model/{id}:
    parameters: [{ $ref: '#/components/parameters/idPath' }]
    get:
      tags: [设备机型]
      summary: 查询设备机型详情
      operationId: getEquipmentModel
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EquipmentModel'
    put:
      tags: [设备机型]
      summary: 更新设备机型
      operationId: updateEquipmentModel
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EquipmentModelCreate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [设备机型]
      summary: 删除设备机型
      operationId: deleteEquipmentModel
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...